+++ CHANGELOG.md
@@ -7,7 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
-// Add your changes here and then delete this line
+### Fixed
+
+- SpotifyException now thrown when a request fails & has no response ( fixes #571, #581 )
 
 ## [2.16.0] - 2020-09-16